Update bb.js
Added conditional arg "violence" to allow the script to be call with a lowered cha requirement for it to start inciting violence.

Changed the assLevel required to trigger while loop in violence() to 5 or less, instead of exactly 0 to combat a bug currently making all bb actions execute in 1 second.