Update warGang.js
Added ability to click the title bar to toggle between full and compact modes (sometimes requires a few clicks for onClick() to trigger)

As a result tem was replaced with raw React.createElement, width and height were turned to let variables instead of const to allow they to be changed after initialization as well a function was added to handle changing the width and height values on a toggle.

Resizing tail now happens of script refresh rate, making manual custom resizing impossible.