Update hashServers.js
Many updates, all the v3.0 stuff with ui namespace

Changed the logic so it now upgrades lower server properties first, by sorting them in least to most and purchasing upgrades we can afford in that order.

Removed unit unit level cap flag and renamed it level. Which will flag on level (old unit level cap x number of servers) when setting level cap on call that is the most a server will be allowed to upgrade to.