Update shotty.js
Grow threads now gets 1 added to it's array empty array 128 times to help with a rare bug when prepping and grow threads array not having enough values stored in it.

Forgot to double weakens ram costs to account for it being used twice when displayed in the debug screen. This change only effects the debug screen display and the rest of the script should still work the same.